Splitting a big test's file into different files to tackle a small part of the tests
